# feat: replace `deep-equal` with `react-fast-compare`