--- a/package.json
+++ b/package.json
@@ -66,8 +66,8 @@
   },
   "dependencies": {
     "@babel/runtime": "^7.9.2",
-    "deep-equal": "^2.0.1",
     "prop-types": "^15.6.1",
+    "react-fast-compare": "^3.0.1",
     "warning": "^4.0.2"
   },
   "devDependencies": {