`react-fast-compare` is more lightweight and faster alternative.

### Demo app bundle size comparison

| Before    | After     | Diff     |
|-----------|-----------|----------|
| 288.67 KB | 255.96 KB | -32.71 KB |

### Performance comparison

[Benchmark](https://github.com/FormidableLabs/react-fast-compare/blob/master/benchmark/index.js) with included `deep-equal` (strict-mode) suite.

```
react-fast-compare x 196,844 ops/sec ±0.41% (92 runs sampled)
fast-deep-equal x 192,322 ops/sec ±0.23% (92 runs sampled)
lodash.isEqual x 34,136 ops/sec ±0.37% (94 runs sampled)
nano-equal x 156,040 ops/sec ±0.56% (90 runs sampled)
shallow-equal-fuzzy x 121,470 ops/sec ±1.42% (93 runs sampled)
deep-equal (strict) x 145 ops/sec ±1.77% (57 runs sampled)
```

196,844 ops/sec vs 145 ops/sec