@khmm12 2020-04-08
Just my opinion and nothing more. 

Personally I would:

1. Replace deep equal comparison with shallow.
2. Add to README a section describing circular updates issue and how to avoid.

by next reasons:

1. Shallow equal comparison for modifiers covers most of use cases.
2. Deep equal comparison solves circular updates issue until modifiers doesn't contain inline functions. That means deep equal didn't solve the issue, but "hidden" for only one case.

```jsx
const [, forceUpdate] = useReducer(v => v + 1, 0)

<Popper
   modifiers={[
     { name: 'boom', enabled: true, phase: 'write', fn: () => forceUpdate() } 
  ]}
>
/* code */
</Popper>
```