@khmm12 2020-04-08
Ooops, I see react-popper is almost rewritten with hooks API and without `deep-equal` usage. So if the PR is not more actual, please close.