@FezVrasta 2020-04-08
Thanks for the PR.

As you noticed, I removed the deep-equal dependency from the new version, but the circular update issue could happen with my pending PR. I think I'd rather defer the responsibility to the consumers, but I need to write some good docs to help consumers understand the issue.