@FezVrasta 2020-04-09
I played a bit more with the new `usePopper` Hook, and I realized it's basically impossible to get it right at first, because as soon you pass an options object, it will start looping.

We should ask the consumers to always use `useMemo` to memoize the options, but it would be a terrible ergonomic... So I ended up adding react-fast-compare to avoid this issue.