@khmm12 2020-04-10
@FezVrasta It's ok (or expected) due to reconciler implementation, because hook arguments are not external dependencies unlike of component props produce causes child component update without affect to external (parent). 

Taking into account that `react-fast-compare` is fast enough to not add extra noticeable overhead, it wouldn't be problem.

> We should ask the consumers to always use useMemo to memoize the options, but it would be a terrible ergonomic... 

IMHO options memoization should be required only if it's under developer responsibility like in circular update issue due to modifiers that produce component update. In other words implementation shouldn't affect to API.  So please don't change your mind, this is right compromise 🙏.